refactor: refactor cjsToEsm.cjs to ES6 notation
