+++ src/cli/env.js
@@ -1,5 +1,6 @@
 const parseEnv = () => {
   const result = [];
+  
   for (const [k, v] of Object.entries(process.env)) {
     if (k.startsWith('RSS_')) {
       result.push(`${k}=${v}`);
